[Refactor] 메인페이지 Footer의 파일명을 FixedFooter로 수정
